Updated Collision - Reflipped Y axis

+++ EdgeLibraryMain/EdgeLibrary/Elements/Sprite/Collision/CollisionBody.cs
@@ -75,7 +75,7 @@ namespace EdgeLibrary
                     return new CollisionBody(new ShapeCircle(sprite.Position, (sprite.Width + sprite.Height) / 4 * ((sprite.Scale.X + sprite.Scale.Y) / 2)), layers); //It's the average over 2, because the average of width+height is the diameter and this is the radius
                     break;
                 case ShapeTypes.rectangle:
-                    return new CollisionBody(new ShapeRectangle(sprite.Position, sprite.Width * (int)sprite.Scale.X, sprite.Height * (int)sprite.Scale.Y), layers);
+                    return new CollisionBody(new ShapeRectangle(sprite.Position, sprite.Width * sprite.Scale.X, sprite.Height * sprite.Scale.Y), layers);
                     break;
             }
             return null;
